correção de bug no sleep.cpp

+++ client/sleep.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main(int argc, char **argv){
     string::size_type sz;
-    this_thread::sleep_for( std::chrono::nanoseconds(stoi(argv[1], &sz)));
+    this_thread::sleep_for( std::chrono::nanoseconds(stol(argv[1], &sz)));
 
     return 0;
 }
